Convert exercises into a controlled component
